+++ game.server/src/handlers/use.card.handler.ts
@@ -5,7 +5,6 @@ import { getGamePacketType } from '../utils/type.converter.js';
 import { gamePackTypeSelect } from '../enums/gamePacketType.js';
 
 import { sendData } from '../utils/send.data.js';
-import { broadcastDataToRoom } from '../utils/notification.util.js';
 
 import { GlobalFailCode } from '../generated/common/enums.js';
 import { CardType } from '../generated/common/enums.js';
@@ -49,52 +48,20 @@ const useCardHandler = async (socket: GameSocket, gamePacket: GamePacket) => {
 	}
 
 	/// 2. 유즈케이스 호출
-	const { useCardResponse, useCardNotification, userUpdateNotification } = await useCardUseCase({
+	const res = await useCardUseCase(
 		userId,
 		roomId,
 		cardType,
 		targetUserId,
-	});
+	);
 
 	/// 3. 유즈케이스 결과에 따라 응답/알림 전송
 	const useCardResponsePacket = createUseCardResponsePacket(
-		useCardResponse.success,
-		useCardResponse.GlobalFailCode,
+		res.success,
+		res.failcode,
 	);
 	sendData(socket, useCardResponsePacket, GamePacketType.useCardResponse);
 
-	if (useCardNotification && useCardNotification.targetUserId) {
-		const useCardNotificationPacket = createUseCardNotificationPacket(
-			useCardNotification.cardType,
-			useCardNotification.userId,
-			useCardNotification.targetUserId,
-		);
-
-		// 장착이 가능한가? equipCard : useCard
-		if (useCardNotification.cardType >= 13 && useCardNotification.cardType <= 20)
-			broadcastDataToRoom(
-				room.users,
-				useCardNotificationPacket,
-				GamePacketType.equipCardNotification,
-			);
-		else
-			broadcastDataToRoom(
-				room.users,
-				useCardNotificationPacket,
-				GamePacketType.useCardNotification,
-			);
-	}
-
-	if (userUpdateNotification) {
-		const userUpdateNotificationPacket = createUserUpdateNotificationPacket(
-			userUpdateNotification.user,
-		);
-		broadcastDataToRoom(
-			room.users,
-			userUpdateNotificationPacket,
-			GamePacketType.userUpdateNotification,
-		);
-	}
 };
 
 /** 오류코드:잘못된요청을 일괄 처리하기 위한 함수 */
@@ -120,38 +87,6 @@ export const createUseCardResponsePacket = (
 	};
 
 	return ResponsePacket;
-};
-
-export const createUseCardNotificationPacket = (
-	cardType: CardType,
-	userId: string,
-	targetUserId: string,
-): GamePacket => {
-	const NotificationPacket: GamePacket = {
-		payload: {
-			oneofKind: GamePacketType.useCardNotification,
-			useCardNotification: {
-				cardType: cardType,
-				userId: userId,
-				targetUserId: targetUserId !== '0' ? targetUserId : '0',
-			},
-		},
-	};
-
-	return NotificationPacket;
-};
-
-export const createUserUpdateNotificationPacket = (user: User[]): GamePacket => {
-	const NotificationPacket: GamePacket = {
-		payload: {
-			oneofKind: GamePacketType.userUpdateNotification,
-			userUpdateNotification: {
-				user: user,
-			},
-		},
-	};
-
-	return NotificationPacket;
 };
 
 export default useCardHandler;
