+++ game.server/src/useCase/use.card/use.card.usecase.ts
@@ -1,43 +1,24 @@
-import { getRoom } from '../../utils/room.utils';
+import { GamePacket } from '../../generated/gamePacket';
+import { GamePacketType } from '../../enums/gamePacketType';
 import { CardType, GlobalFailCode } from '../../generated/common/enums';
 import { User } from '../../models/user.model';
 
+import { getRoom } from '../../utils/room.utils';
+
 import { applyCardEffect } from '../../utils/apply.card.effect';
+import { broadcastDataToRoom } from '../../utils/notification.util.js';
 
-interface UseCardInput {
-	userId: string;
-	roomId: number;
-	cardType: CardType;
-	targetUserId?: string;
-}
-
-interface UseCardOutput {
-	useCardResponse: {
-		success: boolean;
-		GlobalFailCode: GlobalFailCode;
-	};
-	useCardNotification?: {
-		cardType: CardType;
-		userId: string;
-		targetUserId?: string;
-	};
-	userUpdateNotification?: {
-		user: User[];
-	};
-}
-
-export const useCardUseCase = async (input: UseCardInput): Promise<UseCardOutput> => {
-	const { userId, roomId, cardType, targetUserId } = input;
+export const useCardUseCase = async (userId:string, roomId:number, cardType:CardType, targetUserId:string): Promise<{success:boolean, failcode:GlobalFailCode}> => {
 
 	const room = getRoom(roomId);
 	if (!room) {
-		return { useCardResponse: { success: false, GlobalFailCode: GlobalFailCode.ROOM_NOT_FOUND } };
+		return { success: false, failcode: GlobalFailCode.ROOM_NOT_FOUND };
 	}
 
 	// 카드 타입 검증
 	if (cardType === CardType.NONE) {
-		console.warn(`[useCardHandler] 잘못된 카드 타입 요청: NONE`);
-		return { useCardResponse: { success: false, GlobalFailCode: GlobalFailCode.INVALID_REQUEST } };
+		console.error(`[useCardHandler] 잘못된 카드 타입 요청: NONE`);
+		return {  success: false, failcode: GlobalFailCode.INVALID_REQUEST  };
 	}
 
 	console.log(
@@ -46,9 +27,72 @@ export const useCardUseCase = async (input: UseCardInput): Promise<UseCardOutput
 
 	await applyCardEffect(roomId, cardType, userId, targetUserId!);
 
+	// useCardNotification 패킷 전달
+	const useCardNotificationPacket = createUseCardNotificationPacket(
+		cardType,
+		userId,
+		targetUserId,
+	);
+
+	// 장착이 가능한가? equipCard : useCard
+	if (cardType >= 13 && cardType <= 20)
+		broadcastDataToRoom(
+			room.users,
+			useCardNotificationPacket,
+			GamePacketType.equipCardNotification,
+		);
+	else
+		broadcastDataToRoom(
+			room.users,
+			useCardNotificationPacket,
+			GamePacketType.useCardNotification,
+		);
+	
+
+	// userUpdateNotification 패킷 전달
+	const userUpdateNotificationPacket = createUserUpdateNotificationPacket(
+		room.users,
+	);
+	broadcastDataToRoom(
+		room.users,
+		userUpdateNotificationPacket,
+		GamePacketType.userUpdateNotification,
+	);
+
+
 	return {
-		useCardResponse: { success: true, GlobalFailCode: GlobalFailCode.NONE_FAILCODE },
-		useCardNotification: { cardType: cardType, userId: userId, targetUserId: targetUserId },
-		userUpdateNotification: { user: room.users },
+		success: true, failcode: GlobalFailCode.NONE_FAILCODE,
 	};
+};
+
+export const createUseCardNotificationPacket = (
+	cardType: CardType,
+	userId: string,
+	targetUserId: string,
+): GamePacket => {
+	const NotificationPacket: GamePacket = {
+		payload: {
+			oneofKind: GamePacketType.useCardNotification,
+			useCardNotification: {
+				cardType: cardType,
+				userId: userId,
+				targetUserId: targetUserId !== '0' ? targetUserId : '0',
+			},
+		},
+	};
+
+	return NotificationPacket;
+};
+
+export const createUserUpdateNotificationPacket = (user: User[]): GamePacket => {
+	const NotificationPacket: GamePacket = {
+		payload: {
+			oneofKind: GamePacketType.userUpdateNotification,
+			userUpdateNotification: {
+				user: user,
+			},
+		},
+	};
+
+	return NotificationPacket;
 };
